Module Tezos_dac_lib

module Api_version : sig ... end

Api_version module is used for versioning DAC API.

module Certificate_repr : sig ... end

module Dac_clic_helpers : sig ... end

module Dac_plugin : sig ... end

Protocol.Sc_protocol_reveal_hash.t is unknown to modules outside the protocol and only known at runtime. To avoid the proliferation of functors in the dac node, hash hides the dynamic Protocol.Sc_protocol_reveal_hash.t behind an abstract static type. An instance of Dac_plugin.T behaviour of operations on hash.

module Monitor_services : sig ... end

module RPC_services : sig ... end

module Signature_repr : sig ... end