Module Tezos_dac_lib.Certificate_repr

module V0 : sig ... end

This is the first version of the Certificate_repr.t. In the future if it is needed to add a field of modify type of an existing field, one must simply create a V1 module with the new definition of type t.

type t =

| V0 of V0.t

val encoding : t Tezos_base.TzPervasives.Data_encoding.t

Used to return any version of Certificate_repr on DAC RPC endpoints.

val get_root_hash : t -> Dac_plugin.raw_hash

Helper to get root_hash from any given version of Certificate_repr.

val get_aggregate_signature : t -> Tezos_crypto.Aggregate_signature.signature

Helper to get aggregate_signature from any given version of Certificate_repr.

val get_witnesses : t -> Z.t

Helper to get witnesses from any given version of Certificate_repr.

val get_version : t -> int

Helper to get version from any given version of Certificate_repr.

val all_committee_members_have_signed : 'a list -> t -> bool

Certificate_repr.all_committee_members_have_signed committee_members certificate will return true if all members have signed, otherwise it will return false.