Module Certificate_repr.V0

This is the first version of the Certificate_repr.t. In the future if it is needed to add a field of modify type of an existing field, one must simply create a V1 module with the new definition of type t.

type t

Representation of a Data Availibility Committee Certificate. Type is private to make sure correct version is used. Use make function to create a Certificate_repr.V0.t.

val make : Dac_plugin.raw_hash -> Tezos_crypto.Aggregate_signature.signature -> Z.t -> t

Create a Certificate_repr.V0.t from given Dac_plugin.raw_hash, Tezos_crypto.Aggregate_signature.signature and Z.t. This function is in charge to add the correct version.

module Protocol_dependant : sig ... end

All the following functions are related to the Protocol. They are used only by the `command_handlers` module exposing CLI commands for Kernel SDK.