Module V0.Protocol_dependant

All the following functions are related to the Protocol. They are used only by the `command_handlers` module exposing CLI commands for Kernel SDK.

val serialize_certificate : Dac_plugin.hash Tezos_base.TzPervasives.Data_encoding.t -> root_hash:Dac_plugin.hash -> aggregate_signature:Tezos_crypto.Aggregate_signature.signature -> witnesses:Z.t -> Tezos_base.TzPervasives.Bytes.t

Serialize the provided Dac_plugin.hash, Tezos_crypto.Aggregate_signature.signature and Z.t with the correct encoding. Dac_plugin.t is needed to provide the correct encoding related to the correct protocol.