Module Tezos_dac_lib.Monitor_services

module V0 : sig ... end

V0 is experimental DAC API. V0 is deprecated, however for the time being the API will be binding. It will be used by 1M/tps demo. The plan is to remove it once we get rid of the Legacy mode. Use at your own risk!