Module Monitor_services.V0

V0 is experimental DAC API. V0 is deprecated, however for the time being the API will be binding. It will be used by 1M/tps demo. The plan is to remove it once we get rid of the Legacy mode. Use at your own risk!

module S : sig ... end

val root_hashes : Tezos_rpc.Context.streamed -> (Dac_plugin.raw_hash Lwt_stream.t * Tezos_rpc.Context.stopper) Tezos_base.TzPervasives.Error_monad.tzresult Lwt.t

root_hashes streamed_cctxt raw_hash returns a stream of root hashes and a stopper for it.

Stream is produced by calling RPC "GET v0/monitor/root_hashes".

val certificate : Tezos_rpc.Context.streamed -> Dac_plugin.raw_hash -> (Certificate_repr.t Lwt_stream.t * Tezos_rpc.Context.stopper) Tezos_base.TzPervasives.Error_monad.tzresult Lwt.t

certificate streamed_cctxt raw_hash returns a stream and a stopper for monitoring certificate updates for a given root hash.

Stream is produced by calling RPC "GET v0/monitor/certificate".