Module Tezos_dac_lib.RPC_services

module V0 : sig ... end

V0 is experimental DAC API. V0 is deprecated, however for the time being the API will be binding. It will be used by 1M/tps demo. The plan is to remove it once we get rid of the Legacy mode. Use at your own risk!

val get_health_live : ([ `GET ], unit, unit, unit, unit, bool) Tezos_rpc.Service.service

GET dac/health/live returns true if Node_context.get_status cctxt is Starting or Ready.

val get_health_ready : ([ `GET ], unit, unit, unit, unit, bool) Tezos_rpc.Service.service

GET dac/health/ready returns true if Node_context.get_status cctxt is Ready and fail with tzfail Dac_node_not_ready otherwise.

module V1 : sig ... end

V1 is a second major DAC API release. V1 API is work in progress. Do not use!