Module RPC_services.V0

V0 is experimental DAC API. V0 is deprecated, however for the time being the API will be binding. It will be used by 1M/tps demo. The plan is to remove it once we get rid of the Legacy mode. Use at your own risk!

val get_preimage : ([ `GET ], unit, unit * Dac_plugin.raw_hash, unit, unit, Tezos_base.TzPervasives.Bytes.t) Tezos_rpc.Service.service

"GET v0/preimage" requests the preimage of hash, consisting of a single page, from cctxt. When the request succeeds, the raw page will be returned as a sequence of bytes.

val put_dac_member_signature : ([ `PUT ], unit, unit, unit, Signature_repr.t, unit) Tezos_rpc.Service.service

"PUT v0/member_signature" endpoint stores the signature generated from signing hex_root_hash by dac_member_pkh.

val get_certificate : ([ `GET ], unit, unit * Dac_plugin.raw_hash, unit, unit, Certificate_repr.t) Tezos_rpc.Service.service

"GET v0/certificate" endpoint returns the DAC certificate for the provided root_page_hash.

val get_serialized_certificate : ([ `GET ], unit, unit * Dac_plugin.raw_hash, unit, unit, Tezos_base.TzPervasives.String.t option) Tezos_rpc.Service.service

"GET v0/serialized_certificates" endpoint returns the binary encoded DAC certificate for the provided root_page_hash where contained root_hash used encoding is compatible with the Kernel SDK.

val get_missing_page : ([ `GET ], unit, unit * Dac_plugin.raw_hash, unit, unit, Tezos_base.TzPervasives.Bytes.t) Tezos_rpc.Service.service

"GET v0/missing_page/page_hash" Observer fetches the missing page from a Coordinator node. The missing page is then saved to a page store before returning the page as a response.

module Coordinator : sig ... end