Module V0.Coordinator

val post_preimage : ([ `POST ], unit, unit, unit, Tezos_base.TzPervasives.Bytes.t, Dac_plugin.raw_hash) Tezos_rpc.Service.service

"POST v0/preimage" sends a payload to the DAC Coordinator. It returns a hex encoded root page hash, produced by Merkle_tree_V0 pagination scheme. On the backend side it also pushes root page hash of the preimage to all the subscribed DAC Members and Observers.