Module RPC_services.V1

V1 is a second major DAC API release. V1 API is work in progress. Do not use!

val get_pages : ([ `GET ], unit, unit * Dac_plugin.raw_hash, unit, unit, Tezos_base.TzPervasives.Bytes.t) Tezos_rpc.Service.service

"GET v1/pages" requests the preimage of hash, consisting of a single page, from cctxt. When the request succeeds, the raw page will be returned as a sequence of bytes.