Module Tezos_dac_lib.Signature_repr

type t

t is a representation of committee member signature.

val encoding : t Tezos_base.TzPervasives.Data_encoding.t

encoding dac_plugin returns the t Data_encoding.t given a dac_plugin.

val make : Dac_plugin.raw_hash -> Tezos_crypto.Aggregate_signature.signature -> Tezos_crypto.Aggregate_signature.public_key_hash -> t

make root_hash signature pkh creates a t. In addition, it ensures that the underlying Aggregate_signature.t always corresponds to Unknown variant.

Motivation for the Unknown variant: Unknown variant uses "asig" JSON prefix, whereas Bls12_381 variant uses "BLsig" JSON prefix. To ensure consistent signature prefix, when data is exchanged accross the DAC network we use Unknown variant only.

val get_root_hash : t -> Dac_plugin.raw_hash

get_root_hash signature returns a root hash of a given signature representation.

val get_signature : t -> Tezos_crypto.Aggregate_signature.signature

get_signature signature returns an aggregated signature of a given signature representation.

val get_signer_pkh : t -> Tezos_crypto.Aggregate_signature.public_key_hash

get_signer_pkh signature returns a signer public key hash of a given signature representation.